+++ pingouin/plotting.py
@@ -476,7 +476,7 @@ def plot_paired(data=None, dv=None, within=None, subject=None, order=None,
         >>> df = pg.read_dataset('mixed_anova').query("Group == 'Meditation'")
         >>> # df = df.query("Group == 'Meditation' and Subject > 40")
         >>> pg.plot_paired(data=df, dv='Scores', within='Time',
-                           subject='Subject', orient='h')  # doctest: +SKIP
+        ...                subject='Subject', orient='h')  # doctest: +SKIP
 
     With the boxplot on the foreground: